// src/generated/olympusUnits.ts
export type LibChainId = (typeof LibChainId)[keyof typeof LibChainId];

export const LibChainId = {
  NUMBER_1: 1,
  NUMBER_11155111: 11155111,
} as const;

export type SeasonsPositionDirection =
  (typeof SeasonsPositionDirection)[keyof typeof SeasonsPositionDirection];

export const SeasonsPositionDirection = {
  up: "up",
  down: "down",
  none: "none",
  new: "new",
} as const;

export type SeasonsLeaderboardEntry = {
  address: string;
  positionChange: number;
  positionDirection: SeasonsPositionDirection;
  rank: number;
  totalUnits: string;
};